* sdk3.x: (35 commits)
  移除多人会议代码
  update conference xib
  1v1实时通讯  新版UI
  update
  update
  update redpacket message xib
  update project build setting
  update launchimage
  fix searchbar 点击下移
  fix 行缩进混乱
  fix
  去掉旧版的类
  update search controller
  取掉搜索栏
  fix build fail
  update search controller for chatroom list
  删掉没用的类
  update search controller for public group list
  update search controller for group list (bug: search begin, offset 50)
  update contact list
  ...

# Conflicts:
#	ChatDemo-UI3.0/ChatDemo-UI3.0.xcodeproj/project.pbxproj
#	ChatDemo-UI3.0/ChatDemo-UI3.0/Class/Call/Conference/ConferenceViewController.m
#	ChatDemo-UI3.0/ChatDemo-UI3.0/Class/ChatDemoHelper.m

+++ ChatDemo-UI3.0/ChatDemo-UI3.0/Class/Call/EMCallOptions+NSCoding.m
@@ -30,7 +30,7 @@
 {
     [aCoder encodeBool:self.isSendPushIfOffline forKey:@"emIsSendPushIfOffline"];
     [aCoder encodeInteger:self.videoResolution forKey:@"emVideoResolution"];
-    [aCoder encodeInt:self.videoKbps forKey:@"emVideoKbps"];
+    [aCoder encodeInt:(int)self.videoKbps forKey:@"emVideoKbps"];
 }
 
 @end
